Make RCT1 toilets use noCsgImages
This changes the RCT1 toilets to use images from CSG along with having a noCsgImages table to make it more in line with the rest of the RCT1 ride objects as it predates the noCsgImages system, it will use the RCT2 toilets sprites if the user does not have RCT1 linked and this gets rid of the rather dubious unmodified original game sprites from the object's images folder. This should also slightly reduce the file size of the object.

Closes #253