Add RCT1 ice cream stall
Adds the RCT1 ice cream stall as a compat object. Due to the object's RCT2 counterpart having non-remappable sprites in it's dat i've supplied copied of the remappable sprites we use for the RCT2 ice cream stall to be used for the fallback images. I've changed their file names to nocsg_# to make this more clear in the images folder that these are solely for fallbacks.