Add all RCT1 water rides
Adds all RCT1 water ride vehicles as compat objects, currently the river rapids boats use doubled sprites with the RCT2 spinning animation due to the lack of a more flexible spinning animation system. But it doesn't affect the look of them ingame and it should be easy to convert them to a new animation system in the future once it's available.