Add Modern Twister train for LSM LC
This is the first Intamin coaster train to hit OpenRCT2. Originally known as `x7.ride.mega_coaster`, this will need conversion code in the game.
Once we get Intamin's new track style in the game under "Mega coaster" we will need to add another preview image to this ride vehicle and link it to that too.
This vehicle is available on most modern LSM launched coasters (Such as Taron (old track style), Velocicoaster/Pantheon/Toutatis (new track style)) and Mega coasters (Such as Kondaa).

We might get some updates on this vehicle in the future if X7 feels well enough to change some minor things he found while working on the model. This shouldn't impact this release though.

Co-Authored-By: X123M3-256 <X123M3-256@users.noreply.github.com>